SCRD-2839 Use the packaged rabbitmq service file
We don't need to generate a systemd service file for
rabbitmq, the zypper package already has one for us.
However, we do need to extend the epmd.socket unit, which
by default only listens on 127.0.0.1, to also listen
to the management network, otherwise the rabbitmq-server
service will fail to start.

Change-Id: I55b2c8c382ee1f5d6352e6b20ec119b34e85d172