configure TLS/SSL for distribution (bsc#1120681)
TLS/SSL for distribution was disabled in Cloud 8 because of a
bug when the distribution port configuration was not honored.
However, as of version 20.2, we don't encounter this problem
anymore. Therefore, we are re-enabling this feature.

Furthermore, the steps to enable TLS for distribution seem to
have changed as well. See

https://www.rabbitmq.com/clustering-ssl.html

Therefore, we are also updating the playbook accordingly.

Change-Id: Ibd83c36ac6fc2a4e035b52072b8ae5b8fd5badfd