Ensure HA queue sync wait fails (SOC-11083)
We sometimes see the repeat until loop checking that all HA queues
have been sync'd reaching the max iteration count, with some queues
still not sync'd, and continue on, without failing, as though all
queues has successfully sync'd. This is possibly due to an Ansible
1.9 bug in repeat until handling.

To address this add some explicit checking of the registered results
of the action that checks for HA queues being sync'd and fail if the
registered results indicate that the queues haven't in fact sync'd.

Change-Id: Iff9c64ca0cb8191badcceb96cd950cbd456902f3