Set HA policy accordingly (SOC-10317, SOC-11082)
To optimize performance, we need to set the number of mirrors for the
queues to the ceiling of number of cluster nodes divided by 2. However,
we cap the total to no more than 3 mirrors regardless of number of
cluster nodes, per recommendation, as more than 3 is overkill from
cost/benefit standpoint.

We aslo need to mirror the reply_ queues as Neutron services are
depending on them to be available at all times.

Also,in commit d84f662b47e383359c2b8aa1477c66759a733807[1], we
indiscriminately clear the old HA policy and set it to a new one due
to a bug in the rabbitmq_policy Ansible module, that it won't update
the pattern and tags if the policy name already exist. Since the policy
setting logic is exercise in rabbitmq-start.yml playbook, this could
increase disruptions to the system which may lead to some unpredictable
failures. To minimize the impact, we should only set the HA policy if it's not
up-to-date.

1. https://gerrit.prv.suse.net/#/c/ardana/mq-ansible/+/6388/
(cherry-picked from da922ce8f54546ee9083e6900a5d326807685aae)

Change-Id: Ie8e8d714fccf58b4332bb243c5c48304fe53d796