Ensure epmd.service started/stopped independent of rabbitmq (SOC-6780)
Recent update & upgrade testing in virtual and baremetal environments
has hit the issue where the epmd.socket fails to start due to address
already in use errors, leading the rabbitmq-server service to be marked
as failed due to failed dependencies, but it seems to have triggered
an epmd instance to have been started anyway which is listening on port
4369, causing subsequent start attempts to fail, and the epmd instance
is not systemd managed so attempting to stop the epmd socket or service
has no effect.

To attempt to avoid hitting this scenario this patch explicitly starts
the epmd.service before attempting to start rabbitmq-server, and also
stop's it after stopping rabbitmq-server.

Change-Id: I5adb41800568f6b17180818ebdca01dddf1f7a55